working on stage management

+++ PA3.java
@@ -11,8 +11,8 @@ public class PA3
     public static void main(String[] args) 
     {
         //no negative parameters, all ints
-        int m = Integer.parseInt(args[0]);
-        int n = Integer.parseInt(args[1]);
+        int m = Double.parseDouble(args[0]);
+        int n = Double.parseDouble(args[1]);
         int maxQ = Integer.parseInt(args[2]);
 
         //parameter checking
